Nested scope
Stmts returns a type based on if and else stmts. Both if and else stmts must be the same type. Block note type checks the type of if/else stmts and the returns stmt and sets the type of blocknode., func def compares blockNode type with the type declared in the symbol table.